fix(ci): extract the archival sweep so the source ratchet stays put
The archive feature added 137 lines to update_roadmap_progress.ts, taking it
from 1500 to 1584 and putting 84 new lines over the source-size ceiling. Two
gates caught it with one cause: `check_source_size_budget` reported 19511
against a baseline of 19427, and its own test asserts baseline == live total,
which is why Node Tests shard 2/4 failed on both runners while every test the
PR itself touches passed.

Raising the baseline is what that gate calls a defect rather than a fix, so the
sweep spawner and the contract prose it carries move to a sibling module —
`archival_sweep.ts`, the same shape `gate_execute.ts` and `resume_probe.ts`
already use in this directory. `_HERE` resolves to the same directory, so the
lookup of `archive_completed_roadmaps.ts` is byte-identical; verified by calling
run_archival_sweep against a throwaway root and confirming `ran=true` rather
than the silent no-op branch a broken path would have taken.

update_roadmap_progress.ts ends at 1497. Total excess is back to 19427 exactly
— a lowering through extraction, not a shrunk comment.
